Add type_id, source name and type to call query
This table will be the central table for building manifolds. It contains
one column for each argument in each function call.

It does not list the JSON serializing functions (packer/unpacker) or the
Morloc type. Rather, these should be stored in two additional tables
that can be referenced through the `type_id` key and the pair of `lang`
and `lang_type`.

So I am currently looking into a 3 table system. Roughly. There are far
too many columns in this table to reasonably be dealt with as a list of
tuples, so I will build a list of records from it, with one record per
manifold. I'll work on these Haskell representations in the next commit
(probably).