Add argument and data writers
I need to carefully map out when inputs are in natural data structures
and when they are JSON.

A Morloc function has the general form:

```
function := f a_1, ... a_n -> a_out
a_i := A_l | t A
```

Where `A_l` is a language-specific data type and `t A` is a universal
representation of type A (e.g. in JSON). The compiler needs to determine
whether `a_i` is `A_l` or `t A`. If it is `t A`, then the argument needs
to be wrapped in a deserialization function `t A -> A_l`.

The arguments themselves can be name, call, raw data and positional.
These all need to be mapped to `A_l` or `t A`. In a single language
system, then all inputs calls are `A_l` unless they are being passed as
user arguments. Thus `t A` exist only for functions that are not a call
and are exported.

I'll probably just implement this system for now and then work out the
multi-lingual case later. I don't see a fundamental difficulty in the
muli-lingual case, it is just a matter of searching the call graph to
find what is calling what.