Add very rough stub R
I need to propagate the bnd variables from declaration to root.

The rough problem I am trying to work out is how to add all the needed
contextual information into the Manifold records such that just these
records and a map to pack functions is sufficient to build all
manifolds.