Sample Morloc script is now fully functional
A lot of confusion comes from the decision about what ID to use in
identifying the manifolds.

For the sourced, exported functions (i.e. the functions that are
imported from some language and then exported directly without being
composed in Morloc functions) the ID is from:

```
?id rdf:type mlc:export ;
    rdf:value "foo" # or whatever the function name is
```

For all other cases, the ID is the call id from:

```
?id rdf:type mlc:call ;
    ...
```

This includes the exported compositions of form:

```
export foo
foo x y = bar ...
```

Where `foo` is the exported function that immediately calls `bar`. The
nexus will then call the id of `bar`, not the export id of `foo`.

Another point of confusion, is when inputs are unpacked. Everything in
the `foo` (from above) will have `x` and `y` in their scope. But they
may be using `x` and `y` in different languages, so `x` and `y` is
passed through the manifolds as raw JSON and only unpacked when passed
to a terminal function.

TODO:

 * I need a type specific packers. Currently I just call the generic
   packer when returning data from a pool. I need to determine which
   manifolds are only used externally (these are the ones that are called
   from the nexus), determine their output, and then call the appropriate
   packer.

 * Refactor the hell out of my code. I am still not happy with the data
   structures I am using. And there are layers and layers of hacks. I
   am on the mountain, I think, but there is a lot of climbing left.

 * Update README and add an apology for the current SPARQL database
   dependency. Nothing that this will be removed soon.