Convert to golden tests
This massively simplies my test workflows. In the past I had nearly 500
lines of test code which I had to manually update with every change to
the RDF handling.

With the golden tests, I just write morloc files. Then the test suite
compiles them to Turtle. I check over the turtle to make sure it is
reasonable. Then use it as my golden files to compare against in the
future.

If I make major changes, I can just delete all the dated gold and make
more. Of course, there is a bit of a danger here, since it would be
quite easy to test what *is* not what *should be*.

Eventually, once I sort out the SPARQL local server problem, I should
add test for generated code as well.