Remove the composition operator
I found adding handling for the composition operator to be pretty
tricky. Relinking the TopRDF objects is extremely tedious (e.g.
appending a new element). Distinguishing between names and calls with no
arguments is also hard. I will just give up for now. Composition syntax
is not necessary at this stage, it really is just syntactic sugar.

This exposes a weakness in my design, though. Adding what should be a
fairly simple feature to the language proves to be exceedingly
complicated. Perhaps I need to return to building data structures rather
than going straight to RDF. Then I would have a set of structures that
can be converted to and from RDF.

I originally threw out the Haskell data structures in preference of
compiling straight to RDF. But the approach is becoming increasingly
problematic.

A program in the Morloc ecosystem can exist in one of 4 forms:

 1. M - A Morloc script
 2. H - A Haskell representation (the parse product)
 3. R - An RDF graph representation
 4. C - The final executable code

We can write this as a matrix of possible transformations:

   | M | H | R | C |
 M | - | 1 | 2 | 3 |
 H | 4 | - | 5 | 6 |
 R | 7 | 8 | - | 9 |
 C | - | - | - | - |

I won't consider auto-conversions and conversions from generated code to
anything else. That leaves 9 interesting cases.

 1. M->H - This is the standard parse job to a Haskell structure. I
 abandoned the approach in the past since I didn't like the weight of
 all the generated data structures and I wanted to do the heavy lifting
 (the type checking and logic) all logic lang.

 2. M->R - This is my new approach of compiling directly from Morloc
 code to RDF. The problem here is that the Haskell form of RDF is very
 unwieldy. The RDF that is created is sub-optimal, due to the contextual
 limitations of the Haskell parser. These are not real limitations,
 rather I am limited by the weak abstractions the triples provide
 relative to the richer Haskell records.

 3. M->C - Bypassing all intermediate structures and going straight to
 code is almost certainly a bad idea.

 4. H->M - This would basically be a pretty-printer

 5. H->R - Convert Haskell data structures to RDF

 6. H->C - Haskell directly to code. This has proven to be easier than
 going RDF to code.

 7. R->M - Another pretty-printer. In the future, when there are many
 flavours of Morloc script, the R->M code generator could be very
 valuable in translating between dialects.

 8. R->H - This is currently implemented in my code generator. Where I
 build Haskell structures from the results of SPARQL queries.

 9. R->C - I tried this, but it is very ugly. Going from raw SPARQL
 queries to code requires working from very poor tables. It would be
 wonderful if I could somehow write a template language with SPARQL
 plugins for inserting data. The problem though is that the logic can be
 quite involved. Also, a single SPARQL query is rather costly, the need
 to bundle the calls, retrieving many values in tandem, complicates the
 process. There may be an elegant solution to the R->M conversion, but
 for now R->M->C approach seems easiest.

So what is the best strategies? We start with M and want C. The possible
paths are:

 1. M->C - no go, too complicated and we need to do typechecking etc
 2. M->H->C - works, can do type inference and typechecking in Hask
 3. M->R->C - may work, type inference and checking with OWL
 4. M->R->H->C - my current approach, pretty convoluted, though it
 breaks down into the steps (M->R) and (R->H->C) with the R stored in a
 SPARQL database.
 5. M->H->R->H->C - The direction I am heading. Where H is a data
 structure that can be converted to and from RDF.

R allows programs to be stored and merged. It also offers the
possibility of elegant declarative type checking and validation. It is
the semantic heart of Morloc. The RDF representation is the true form of
a Morloc program.

The M->R makes for an awful parser. R->C makes for an awful code
generator. So perhaps the M->H->R->H->C model is the only viable one.

OK, so I need to develop a really good Haskell representation of a
Morloc program. It needs to be symmetric: easily parsed to and from R,
from M and to C.