Refactor files and generate stub python code
The generated code is total rubbish, being still a mix of R and python.
And the template file is almost a direct copy of the R one. I will need
to factor out all the common logic. But first I want to get the
generated code roughly working.

I also changed the file structure. Now all language-specific handling is
in `Template` folders.