Added back language grammars
The new grammars are functions of Doc objects. I currently use this
grammar:

```
data Grammar = Grammar {
      gCall     :: Doc -> [Doc] -> Doc
    , gFunction :: Doc -> [Doc] -> Doc -> Doc
    , gComment  :: Doc -> Doc
    , gReturn   :: Doc -> Doc
    , gQuote    :: Doc -> Doc
    , gSource   :: Doc -> Doc
    , gList     :: [Doc] -> Doc
    , gTuple    :: [Doc] -> Doc
    , gRecord   :: [(Doc,Doc)] -> Doc
    , gTrue     :: Doc
    , gFalse    :: Doc
  }
```

R and python are similar enough to allow almost complete code reuse
between the languages. The system will need to be generalized somewhat
in the future when I add in Bash (which barely qualifies as a language)
and typed languages.

I may split `Common` into several sub-commons for different kinds of
languages. We'll see how it goes.

I still need to add trans manifolds.