Purge Walker and mtype to JSON (aeson) handling
There was a lot of good material here that I may want to resurrect in
the future. But all of this was based on my old system and if I do bring
it back it will need to be changed significantly. Probably it will be
easier to just rewrite it from scratch.