Reverse the order of elements
Previously nodes in the RDF graph where linked to their children as
follows:

```
?child0 rdf:_0 ?parent
?child1 rdf:_1 ?parent
```

However there are two problems with this. First, everywhere else in the
graph the direction tends to be parent to child. Second, the subject
node must be a URI and a parent is necessarily a URI, whereas a child
may just as well be a plain value.

Anyway, now I have reversed the order, e.g.:

```
?parent rdf:_0 ?child0
?parent rdf:_1 ?child1
```

I've also made a few other small changes. In records, the recordEntry's
are now also listed as types and their elements are enumerated as lists,
rather than lhs/rhs. This is all part of an effort to simplify the
graph, making it more suitable for access. More on this soon.

P.S. Oh, and the turtle is every so much more readable now!