Swinging a little ways back to a grammar approach
The program is still broken, but it is almost there. The main problem
(that I know of) is that the cis manifolds need to cast their arguments
if they are coming in as JSON. They will be JSON if they either come
from the user or from a foreign call.

Currently all bound arguments come from the use, though this will change
once I add support for high-order functions.