Handle each input with its own assignment
Now inside each manifold, the first thing done is to assign each
manifold input to a local value, e.g.:

```
m85 <- function(n,d){
    a0 <- .morloc_unpack(unpackgeneric
                        ,n
                        ,.name="m85"
                        ,.pool="pool.r")
    a1 <- 0.0
    a2 <- .morloc_unpack(unpackgeneric
                        ,d
                        ,.name="m85"
                        ,.pool="pool.r")
    runif(a0,a1,a2)
}
```

Each input is unpacked, if needed, and assigned to a temporary variable.
Later on assertions will be added as functions of the `a*` variables and
the output. First though, I want to add a type comment before each of
these.