Fixed types in manifolds
I can now print some semblance of a type signature in the pools. It
isn't right on several counts. The formatting is ugly. I have extra
parentheses. More importantly, the MDeclType's are nesting. That is ugly
as hell. I am not really sure whether I should include the type
declarations at all in the type tree. They just complicate things. All
the real info is in the nested function signature.

Yes indeed, I should dump them totally. I will need to reorganize a bit
of info. The MFuncType will need to contain the property list.

I need to test over all this. My code is becoming a little hard to
maintain. The Haskell part is easier, it is all the seams with other
languages that is a pain.