Tried HSparql DSL
Originally I decided to use SPARQl with a little quasiquotation magic
since it was simpler and more elegant than using a Haskell DSL (i.e.
HSparql).

However, my literal SPARQL queries have since proven to be a major
source of bugs. The problem is mostly that they are not typechecked and
it is VERY easy to have small typos that mysteriously lead to weird
results.

So I decided to try the DSL. It worked beautifully for the simple MData
case. However, some of the functionallity of SPARQL is not yet
implemented in HSparql. The library is not under active development, so
this situation is unlikely to change unless I start hacking on it (and I
don't really have the time). The main thing that is missing is the
GROUP_CONCAT aggragating command. I played around with implementing this
in a fork of Rob Stewart's repo. But so far have not succeeded. Even if
I do, it will be awhile before the fork is accepted.

All that to say, the DSL will be the way to go in the future, but for
now I will stick with literals.