Convert Manifold SPARQL query to hsparql
There is a lot of ugly repetitive code in the hsparql queries. You have
define each term before you use it, for example. Most of it is pretty
reasonable, but still, the `hsparql` code is twice the length of the raw
SPARQL code.

Still, the advantages outweight the disadvantages. With the DSL I can
use all my algebraic nodes. And I can work on factoring out common
pieces of coding.