Add simle type checker
The typechecker only checks the case where an argument with a concrete
type receives input from a function with a concrete output. In this
case, the typechecker raises a warning (but does not die) if the types
are not identical.

This is, of course, a far cry from being a "real" typechecker. There is
no handling of constraints and generic types. Also, there is no
searching for potential casting functions.

I think to think through how the typechecker works with the inference
machine. They should be modular, but there is some overlap. And there is
a third machine.

Here's an outline of how the whole process should look:

 1. Resolve generic cases where possible, raising warnings/errors for
    any ambiguous types. If the generics can not resolve to concrete
    cases, they should at least resolve to equivalence classes of
    generics (e.g. where they all must be the same, and are resolved at
    runtime, though I haven't really figured out how that would work).

 2. Check input/output type correctness. Keep track of all conflicts.
    (This is actually pretty trivial after the work of the previous step)

 3. Resolve conflicts with cast paths. These need to be stored somewhere
    and passed along to the code generators.