Add RdfLike, SparqlLike, and SparqlDatbaseLike
This is a first step towards generalizing SPARQL querying beyond just
using a SPARQL endpoint. The first goal along these lines is replacing
the requirement for a SPARQL server with a local call to the Jena `arq`
command, which takes a RDF file and SPARQL file and returns a table of
results.

I wrote RdfLike and SparqlDatabaseLike instances for RDF. These will let
me to run SPARQL queries without the server, I just need to do a little
more wiring.

Next thing is to add a SparqlLike instance for `Query a`.