Fixed arq TSV parsing
Everything was dying because arq was wrapping UNode ids in angle
brackets and quoting fields. Adding unquoting and unangling to the TSV
parser fixed the problem. The solution is a bit hacky, and doesn't
handle escaped values, but it works for now.

The compilation process is terribly slowed down by the multiple calls to
`arq`, which is slow since each call requires firing up Java. This is a
temporary inconvenience, though, since `arq` most certainly is not what
I will be using in the future.