Add documentation and clean Graph* objects
There were a couple un-used GraphPredict and GraphObject constructors. I
removed these. I also added documentation that clarifies the roles of
each of these predicates and objects.

Perhaps I should move all class instances into the `Morloc.Type` module?
That would make importing instances easier. On the downside, it clutters
the module's code. However, the code is already long enough that
searching it is awkward. The human reader would probably be exploring
through the Haddock-produced rich-format pages, which are linked
straight to the source code.

PS - I need to keep any real programming in the dev branch. It is fine
to do some documentation here on master, but any real changes to the
code should be in dev.