Remove play.rq
This was a file of SPARQL queries that I experimented with early in the
transition to SPARQL. I don't need these anymore, so I'll delete them.

I will keep the `rules.ttl` file for now. This file is not currently
used, but represents my tinkering with expressing constraints and checks
on the language through its type system.