Add stub YAML config reader
Currently I don't do anything with the config reader and I use the
defaults for morloc home and library. The next step is to read in the
config early and pass it everywhere it is needed. I am not sure of the
best way to do this. Perhaps build a wrapper for computation, that
integrates the error handling and config passing. Or make a monad stack,
with a Reader monad (for the config) and an Either for error handling.

Unrelated, but my satisfaction with the state of the codebase is
lessening.