Write type signature for a core monad stack
This stack consists of ReaderT, ExceptT, WriterT, StateT and IO monads.
It will allow a configuration to be threaded through the program
(`Reader Config`), errors to be logged (`Either MorlocError`), log
messages to be stored, and state to be stored and accessed, and finally
IO operations to be made.

TODO:
 [ ] choose log and state types
 [ ] formalize log messages and config and state access with sum types
 [ ] actually thread this damn thing through the system