Fix RDF path mangling for ID generation
I am prefixing new IDs with the path to the source code that generated
them. This is kludgy, but for now I am removing the library path (which
is user specific) and replacing the illegal ('/' characters with '__').

I've found two problems with my generated code.

 1. not all of the manifolds are being generated, for example, in
 `sample.loc` only one of the three wrapped functions appears.

 2. The generated pool code do not have the correct paths to source
 code. For example, R is looking in "$MORLOC_LIB/rbase.R" instead of
 "$MORLOC_LIB/rbase/main.R".

I'm not sure of the cause of the first problem. Probably a simple coding
mistake in the generator ...

The second problem, though, is much deeper. I need to maintain a map of
dependencies in the MorlocMonad state. This will contain links from
imports to files, a list of required executables, etc.