Fix bug in MType.hsparql and export all hsparqls
The SPARQL queries are the most bug-prone objects in the source code. I
now export them all so that they can be publicly documented and exposed
to the test suite.

While carefully reading through the MType.hsparql query, I found a
rather serious error. The lang, typename, and property values were all
not being exported. The fact that the properties were not being returned
means that code downstream did not have access to the "packs" and
"unpacks" function properties. Something is still fishy, I'll need to
look into this ...