Begin working on sample-complex.loc
It compiles, but the generated code is broken in many ways.

First, there is this monstrosity at the beginning of `pool.py`

``` python
 #!/usr/bin/env python

 import sys
 import subprocess
 import json

 sys.path.append('/home/z/.morloc/lib')from /home.z..morloc.lib.pybase.pybase import *
```

Then there are the awful manifold namings, for example:

``` python
 # cis manifold
 # len :: (a -> Int)
 def msample-complex.loc_29():
     a0 = xs
     return(len(a0))
```

To fix all this, I will first go through the file by hand and craft it
into something that works. Then I will mess with the generator until I
get that thing.