Fix the python import problems.
`sample-complex.loc` is roughly working now. The return types seem a
little wrong still. R returns everything as a vector. This is reasonable
considering R's data model, but I should be able to over-ride this with
more specfic types. Specifically, the `sum` function should return a Num
not a `[Num]`.