Begin adding C compatibility
Previously, the code for each pool was simply written to a file. This
makes sense for simple interpreted programs. But for compiled programs
(or programs with additional packaging) we need a build step.

This commit introduces a new module `Build.hs`, which includes a `build`
function that dispatches incoming Script objects by language to the
appropriate build function. The interpreted scripts are treated as
before, except that, for now, all are made executable. C programs are
dispatched to a currently undefined `buildC` function.

Next steps:
    1. Define the `buildC` function. For simple cases, this just
       requires calling gcc on the given script.
    2. Add handling to nexus for calling compiled executables. Currently
       the nexus calls interpreted programs using the appropriate command,
       rather than relying on the shebang of an executable. I may switch
       over to an all-shebang approach which would allow the nexus to treat
       compiled executables and interpreted programs in the same way.
    3. Define the `Pools.Template.C::generate` function. I will need to
       generate a main function, an argument parser, etc. I can start
       with shamelessly hard-coding the script to work in a particular
       case. This will allow me to test out the steps 1 and 2 before
       tackling the full detail of the code generation.
    4. Generalize the `generate` function.
    5. Add fortran. This will follow a fairly similar pattern as C, but
       will force me to generalize over compiled languages.