Made stub C functions
There is a lot of refactoring I need to do before continuing with the C
handling code:

  [ ] Replace name mangled manifold calls with manifold numbers. This
      will allow case-switch index access dispatch on ID.
  [ ] Replace the string language names with a sum type with optional
      version info.
  [ ] Replace program string name with a function of the executable
      (or script) file.
  [ ] Add GCC flags to package yaml config files and add parser for
      package config.

I've added stub handling for C. I can generate a nexus and an
executable. But the wiring from nexus to executable is broken as hell.
To see what is going on run `morloc make trig.loc` in the example
folder. Also, note that the `trig.loc` file is actually ignored, since
the `generate` function for C hardcodes a particular C program.