Convert to integer indices
Now the nexus dispatches to pools with the manifold integer ID, rather
than the name. This allows simple case switches, avoiding eval
statements and/or string comparisons (which are awkward in some
languages).

So now our TODO is:
 [x] use integer IDs for manifolds
 [x] use sum type for languages
 [ ] replace program string name with a function of the file.
 [ ] make package YAML parser
 [ ] add GCC flags to package yaml config files

Full C support is still a ways away. Next up is item 3.