Add github/semantic dependency
I don't actually use this package yet, but I will eventually. `semantic`
is designed to be able to parse many langugages into an AST, also to
diff and compare them. This is highly valuable to `morloc`, since it at
least partially removes the blackbox status of source code functions. It
also will be highly valuable for code curation. For example:
  * calculating the true complexity of a module/function
  * determining if a function has truly changed (not just comments)
  * limited type checking on incoming code
  * ensuring the exported functions are actually defined
  * possibly translating between languages