Filled out C grammar (with FIXMES)
Here are the main challenges to generating C code:
 1. types in assignments and prototypes
 2. memory management - allocation (which is type dependent) and memory freeing
 3. compilation - which may require custom compiler flags
 4. pointers and casting

A function declaration now requires the return type, the type of each
argument and the creation of a prototype elsewhere in the code.

I think I will for now abandon my attempt to use the same Grammar record
for R/Python and C. Instead I will work on an independent C generator
and then abstract it to work on other typed, non-garbage collecting,
languages (e.g., fortran).