Factored common Manifold handling out of Common
This is a necessary first step towards generalizing supported languages
across more complex and variable languages. The functions in Manifold.hs
should be retrieve data from manifolds and manifold lists that will be
needed for any form of generator. The remaining functions in "Common.hs"
(a poorly named file) are those that are specific to a particular sort
of language (R, python and perhaps other interpreted, dynamically typed
languages).