Nearly reachieve past functionallity
Somewhere many commits back I broke the foreign call system (among other
things). This is a failure in testing.

I have fixed most of the problems. The remaining one appears in the
`test` composition in the `sample-complex.loc` script. In the R.pool,
the `roll` morloc function (which is the name of the composition) is
called, but this function is not defined in R, and instead the first
function in the composition should be called. I addressed this specific
problem a long time ago. It is disturbing that it has come back.

I need to make a better test suite.