Fix the composition problem
A morloc function is a composition of other morloc functions and source
functions. For example:

```
from "math" import (sqrt, pow)

foo x = sqrt(pow x 2)
```

Here `foo` is a morloc function. When a pool is generated, the function
`foo` cannot be called directly, since it is just an abstraction and is
not defined in any source language. Instead, the `foo` manifold should
return the result of evaluating `sqrt` on `pow x 2`. The current SPARQL
queries identify manifolds that are the top-level in a composition (this
is stored in the Manifold object). To link the `foo` return function to
the appropriate source function `sqrt`, I now build a map from
composition name (e.g., "foo") to manifold (e.g., the `sqrt` Manifold
object). This map is used in Common.hs now to resolve the return calls
from compositions.