Generalize function generators
In the generator Grammar, I've replaced the old `gFuntion`:

```
gFunction :: Doc -> [Doc] -> Doc -> Doc
```

With:

```
gFunction :: GeneralFunction -> Doc
```

The GeneralFunction contains all the same information as the 3 arguments
to the previous function. In addition it stores return type and the type
of each function argument. Further, fields can be added to the record
without breaking existing language generators.

This is a first step towards unifying the R/Python code with the C code.

My approach to unification is to incrementally replace the fields of the
Grammar record with gneralized functions that handle types. Once this is
done, I can refactor the C.hs file to use Grammar and Common.hs
functions.