Allow complex properties
This is a big commit with several changes:

First of all I added handling for specifying function properties with
multiple terms, for example:

```
add :: Num a => a -> a
```

Where `Num a` has two terms. This allos general relations to be
specified for a function. The properties are stored as `[[Text]]`.

On the housekeeping side, I added tests for the new parsing and also
made the OverloadedStrings and FlexibleContexts extensions default.

Everything basically seems to work, however, the `sample-complex.loc`
script seems to fall into an infinite loop. I'm not sure why ...