I um, fixed it?
It compiles, but I am not happy. I guess this will work if we assume
converting [a] into "std:vector<a>" is always the right thing to do. For
tuples I am also converting to vector, which is a shadier action. As for
records ... I structure I guess. Or a class? Well, essentially, there
are more than one way to do it, and I don't capture all of this.