Detect bound variables in the parser
A new triple, with predicate mlc:is_bound, stores a boolean describing
whether or not a variable in a datadeclaration is bound. Now unbound,
named arguments will emit manifolds, though currently they are not used
for anything. The `arithmetic.loc` script tests a map function, but
currently fails. Making this example work is my next priority.