Found a bug, make `arithmetic-c.loc` to see
The problem here is with deserializing the output of the generic `map`
function. The return type is dependent on the type of the wrapped
function. I will need real type inference to fix this problem. For now,
it works with R and python, in at least some cases, since the generic
serialization functions will handle simple cases (primitives and lists
of primitives).