Still fail on pool generation
The MTypeMeta is differing greatly from what it used to be. Compare the
current output with the logs in .morloc/tmp. I really should just
through out all the old structures for type representation. But for now,
I just need to figure out what the getPackers and getUnpackers functions
in Manifold.hs want. Currently the pool generator dies at Common.hs:314
with a "No unpacker found for argument: <>" error.