Add include compiler flags
Automated building: this is going to be so much fun.

Previously in the C/C++ programs I was generating absolute include
paths, for example:

``` C
 #include "/home/bob/.morloc/lib/stuff/foo.h"
```

This is not pretty, but it is specific and avoids searching through too
many directories (where there may be name conflicts).

A prettier solution is to write include statemetns in the compilation
step. Then the code is simply:

``` C
 #include "foo.h"
```

And the generated build command is:

``` sh
g++ ... -I/home/bob/.morloc/lib/stuff
```

I am now storing the module directories in the `Script` objects. They
can be accessed in `Build.hs` to create the `-I` flags.

We need a more permanent solution to our problem.