Considering realizations 2
I need to work through a few examples by hand to divine an algorithm for
inferring concrete types. Should the concrete inferences be done at the
same time as the abstract realizations? Or should it be done in a second
pass? Or should it not be done at all until the manifold stage?

What information is needed to infer concrete types?