(21 tests pass) found location of infinite loop
The expression:

```
f :: forall a . a -> a; f 42
```

passes without an infinite loop, but generates an incorrect result:

```
Main {
    f :: forall a . a -> a
    ((f :: (forall a . a -> a -> Num)) 42.0 :: (Num))
                            --------
}
```

Notice that 1) the generic variables were not resolved to `Num` and 2)
an extra argument was added to f.

An infinite loop (or at least a memory explosion) occurs with the
following expression:

```
f :: forall a . a -> a -> a; f 1 2
```

This is probably a logic bug in `-->E`.