(16/113 fail) Fix primitive concrete handling
Previously everything simply blew up when the typechecker attempted to
infer the type of a concrete primitive. This is of course the wrong
behaviour, since primitives should be expressible in all languages. So
instead I now return a fresh existential type in the given language.

I should perhaps also store the additional constraint that the
existential type must be homologous to the given primative type.