(4/114 fail) fix typing of functional arguments
Currently, when an argument is passed to a function, the variable bound
to the argument is recast throughout the body as a single existential
variable.  This means that a single type will be inferred for the
variable, the same type everywhere the term appears. In this sense, the
variable bound under the lambda is not polymorphic.

I need to test the handling with multiple languages. The variable should
be able to work in different languages.