Partial solution to concrete parameterization
For a long time I have been worrying about how to handle parameterized
types in concrete languages. For example, the most natural default
representation for a list of integers in C++ is `vector<int>`, likewise
for a list of strings `vector<string>`. But we should not have to write
a new instance of a casting function for vector paired with every
possible type.

A simple solution to this problem, that at very least works for the C++
issues that are my immediate concern, is to use parameterized types
where the parameters are both independent concrete types as well as
patterns that are copied into the main type. For example, `vector<int>`
could be encoded as:

``` haskell
    ArrT (TV (Just CppLang) "vector<$1>") [VarT (Just CppLang) "int"]
```

Or written in Morloc script as simply: `"vector<$1>" int`. This
corresponds one-to-one with the Morloc expression `List Int`. The
Haskell encoding would be `[$1] Int` or `[$1] Integer`. The `$1`
indicates a location where the first parameter will be substituted into
the 0th expression. Of course there could be more parameters, for
example `pair<$1,$2>` (C++) or `Map $1 $2` (Haskell).

With this syntax, the `map` function in C++ can be written as:

```
map Cpp :: forall a b . (a -> b) -> "vector<$1>" a -> "vector<$1>" b
```

The polymorphic variables will all be solved in the type inference step.
For example, in:

```
sqrt Cpp :: double -> double
f = map sqrt [1.1, 2.2, 3.3]
```

The terms in the second line will evaluate to:

```
f Cpp :: "vector<$1>" double
map Cpp :: (double -> double) ("vector<$1>" double) ("vector<$1>" double)
sqrt Cpp :: double -> double
[1.1, 2.2, 3.3] :: "vector<$1>" double
```