(1/115 fails) test concrete lambda type inference
I should be able to infer concrete types for delarations. For example:

```
sqrt :: Num -> Num
sqrt R :: numeric -> numeric

foo x = sqrt x
```

Here it obviously should not be necessary to write a signature for
`foo`. And it most certainly should not be necessary to write one for
every possible realization of `sqrt`.