(1/122 fail) fix concrete parameterized existentials
Type inference now mostly works with parameterized concrete types.
Concrete containers will be first cast as parameterized existential
types. The type is indexed by the existential name, say <t0>, which will
eventually be solved as to a concrete type, say "std::vector<$1>". The
parameters will be solved as new information is observed.

The tuple construction is now correctly typed, but I haven't yet
extended this to lists and records. I'll do this next. I will need to
add more tests for these containers that include non-primitices and type
variables.

The final failing case also needs to be addressed, of course.