(2/123 fail) add simpler failing test
The following script should return `forall a@RLang . a@Rlang`, but
instead it incorrectly returns `b@CLang`:

```
foo r :: forall a . (a -> a) -> a -> a;
f c :: b -> b;
foo f 1
> b@CLang
```

Dropping the forall produces the correct `a@RLang` output:

```
foo r :: (a -> a) -> a -> a;
f c :: b -> b;
foo f 1
> a@RLang
```

This seems to be a subtyping issue. What exactly does a concrete
`forall` mean? The problem seems to be that

```
(forall a@R . a@R <: b@C |- b@C)
```

That is, the concrete forall is allowing transitions between languages.
This should instead subtype as follows:

```
(forall a@R . a@R) <: b@C |- (forall a@R . a@R)
```

Subtyping against a foreign langauge tells us nothing about the concrete
type. Well, that is only partially true, it should be generate a
serialization constraint.