Make infer LamE return LamE and add tests
 * added a new kind of tests, `assertTerminalExpr` that removes all type
 annotations from the returned expression and checks the terminal
 expression against the given expression.

 * renamed `exprTestGood` to `assertTerminalType`

`infer` on the LamE pattern was previously returning the body of the
lambda, rather than the entire lambda expression. This screwed up the
build process downstream (besides breaking the rule that `infer` returns
the type of the input expression and the exact input expression
unchanged except for with added types).

The build still does not work, but I'm getting closer. Hopefully I can
have the golden tests online soon ...