Fix generator, add output golden test
`partial-01.loc` now compiles and runs correctly. I added a golden test
for its output. I think a good general approach for golden tests is to
use makefiles and `exp.txt` and `obs.txt` files. This way a simple
function can run every golden test with only the test message and
directory changing. The details will all be in the code and Makefile.