Clean up C++ gen and properly parse modules
The q.loc example still does not work. The C++ code is messy in a lot of
areas (unnecessary declarations, passed parameters, etc). More
importantly, the code does not compile since currying is not handled
properly.

I need to develop a means of currying functions in each supported
language. This is tricky, since currying is not natural in some
languages, where passing functions is awkward (looking at you, C).  It
does appear to be possible in C++ with std::bind.

C is tricky. The natural way to pass a function in C is using a function
pointer. The pointer is the address of an executable block of memory.
The problem is that there is no real way to store state in the function
(apart from adding static variables, which would not play well in a
multi-threaded environment). If we can't pass functions with state, then
we can't use closures or partial function application. This may be a
fundamental limitation to the use of C functions in morloc: they cannot
take functional arguments.

In C++, function arguments will have to be defined with templates. This
is the natural way to do it in C++, so is not an onerous requirement to
set. `std::bind` returns a function object that can be passed readily as
an argument to functions that take a generic function argument. No
problem.

So I need to make one more special case in Common.hs where partial
application is recognized and triggers the proper "partial" grammar
call for the given language.