Remove path from etype
EType stands for 'extended type'. The idea was to append types with
extra information, like properties and constraints. I don't know why I
was also putting path information in. The only case where a source path
should be coupled to a type is for sourced functions. But the source has
nothing to do with the type itself (e.g., where is Int imported from?),
so it makes more sense, and simplifies the code, to move the source
information elsewhere. Now sources are stored in a list of Source
objects in the Module object.